something

+++ src/main/java/by/sviryd/engvoc/controller/SignController.java
@@ -1,16 +1,18 @@
 package by.sviryd.engvoc.controller;
 
-import by.sviryd.engvoc.config.RegistrationTokenExpirationDTConfig;
 import by.sviryd.engvoc.domain.User;
 import by.sviryd.engvoc.domain.VerificationToken;
-import by.sviryd.engvoc.service.*;
+import by.sviryd.engvoc.service.SignUpUserService;
+import by.sviryd.engvoc.service.UserService;
+import by.sviryd.engvoc.service.VerificationTokenService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.View;
 
@@ -22,19 +24,11 @@ import java.util.UUID;
 @RequestMapping("/sign")
 public class SignController {
     @Autowired
-    public RegisterUserService registerUserService;
+    public SignUpUserService signUpUserService;
     @Autowired
     private UserService userService;
     @Autowired
     private VerificationTokenService verificationTokenService;
-    @Value("${info.url}")
-    private String url;
-    @Autowired
-    private RegistrationTokenExpirationDTConfig registrationTokenExpirationDTConfig;
-    @Autowired
-    private LocaleExceptionAttributeWrapperService localeExceptionAttributeWrapperService;
-    @Autowired
-    private MessageI18nService messageI18nService;
 
     @GetMapping("/up")
     public String up() {
@@ -63,16 +57,35 @@ public class SignController {
     }
 
     @GetMapping("/activate/{token}")
-    public String activate(Model model, @PathVariable VerificationToken token) {
+    public String activate(@PathVariable VerificationToken token) {
         if (token != null) {
-            User user = token.getUser();
-            user.setToken("1Aa".concat(UUID.randomUUID().toString()));
-            user.setActive(true);
-            userService.save(user);
-            verificationTokenService.delete(token);
-            return "redirect:/registration/activationSuccess";
+            if (token.isExpiredToken()) {
+                return "redirect:" + "/sign/activation/expiredToken";
+            } else {
+                User user = token.getUser();
+                user.setToken("1Aa".concat(UUID.randomUUID().toString()));
+                user.setActive(true);
+                userService.save(user);
+                verificationTokenService.delete(token);
+                return "redirect:" + "/sign/activation/success";
+            }
+        } else {
+            return "redirect:" + "/sign/activation/failure";
         }
-        model.addAttribute("activationError", "Введен неверный код!");
-        return "activation";
+    }
+
+    @GetMapping("/activation/expiredToken")
+    public String activationExpiredToken() {
+        return "main";
+    }
+
+    @GetMapping("/activation/success")
+    public String activationSuccess() {
+        return "main";
+    }
+
+    @GetMapping("/activation/failure")
+    public String activationFailure() {
+        return "main";
     }
 }
